+++ internal/importing/cgofixture/cgo.go
@@ -1,0 +1,10 @@
+package cgofixture
+
+/*
+#include <stdlib.h>
+*/
+import "C"
+
+func Value() int {
+	return 1
+}